chore(web): lint
