+++ kagami/webui/src/index.tsx
@@ -1,12 +1,11 @@
-//tsx:typescript + jsx 虚拟DOM
+// Tsx:typescript + jsx 虚拟DOM
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './global.scss';
-import App from './App';
-
+import App from './App.js';
 
 const root = ReactDOM.createRoot(
-  document.getElementById('root') as HTMLElement
+  document.querySelector('#root')!,
 );
 root.render(<App />);
