+++ kagami/webui/src/page/Ad/Announcement/index.tsx
@@ -1,17 +1,16 @@
-import React, { useEffect, useState } from 'react';
-import styles from './styles.module.scss';
+import React, {useEffect, useState} from 'react';
 import MarkdownIt from 'markdown-it';
+import styles from './styles.module.scss';
 // 安装 @types/markdown-it 后，可以导入类型
 
-
-interface Announcement {
+type Announcement = {
   id: number;
   data: {
     title: string;
     content: string;
     date: string;
   };
-}
+};
 
 const md = new MarkdownIt();
 
@@ -21,9 +20,13 @@ const AdSection: React.FC = () => {
   useEffect(() => {
     // 从后端获取公告数据
     fetch('http://127.0.0.1:4523/m1/5454758-5129914-default/api/announcement')
-      .then(response => response.json())
-      .then(data => setAnnouncement(data))
-      .catch(error => console.error('获取公告失败:', error));
+      .then(async response => response.json())
+      .then(data => {
+        setAnnouncement(data as Announcement[]);
+      })
+      .catch((error: unknown) => {
+        console.error('获取公告失败:', error);
+      });
   }, []); // 依赖项为空，只在组件挂载时执行一次
 
   return (
