+++ kagami/webui/src/page/IntroductionPage/FirstSection/styles.module.scss
