+++ kagami/webui/src/page/IntroductionPage/SecondSection/styles.module.scss
