--- kagami/webui/src/page/Mirror.tsx
@@ -1,19 +1,0 @@
-import React from "react";
-import { Link,useNavigate } from "react-router-dom";
-import Firstsection from '../introduction_page/Firstsection';
-
-
-
-
-
-
-const Mirrorpage=()=>{
-
-return(
-    <div className="Advertisementpage">
-      <Firstsection />
-
-    </div>
-)
-}
-export default Mirrorpage;
