+++ kagami/webui/src/page/Mirror/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import {Link, useNavigate} from 'react-router-dom';
+import FirstSection from '../IntroductionPage/FirstSection';
+
+const Mirrorpage = () => (
+  <div className='Advertisementpage'>
+    <FirstSection />
+  </div>
+);
+
+export default Mirrorpage;
