--- kagami/webui/src/page/State.tsx
@@ -1,19 +1,0 @@
-import React from "react";
-import { Link,useNavigate } from "react-router-dom";
-import Firstsection from '../introduction_page/Firstsection';
-
-
-
-
-
-
-const Statepage=()=>{
-
-return(
-    <div className="Advertisementpage">
-      <Firstsection />
-
-    </div>
-)
-}
-export default Statepage;
