+++ kagami/webui/src/page/State/state.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import {Link, useNavigate} from 'react-router-dom';
+import FirstSection from '../IntroductionPage/FirstSection';
+
+const Statepage = () => (
+  <div className='Advertisementpage'>
+    <FirstSection />
+  </div>
+);
+
+export default Statepage;
